Extract `FeedImageDataLoader` from view controller file. Also, arrange all file into respective folder structures
